feat-#16: move typeorm repos to infrastructure
