+++ src/item/application/queries/handlers/get-items.handler.ts
@@ -2,7 +2,7 @@ import { IQueryHandler, QueryHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { ItemEntity } from '../../../entities';
-import { ItemReadRepository } from '../../../repositories';
+import { ItemReadRepository } from '../../../infrastructure/repositories';
 import { RpcExceptionService } from '../../../../utils';
 import { GetItemsQuery } from '../impl';
