+++ src/item/infrastructure/infrastructure.module.ts
@@ -1,4 +1,11 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { ItemReadRepository, ItemWriteRepository } from './repositories';
 
-@Module({})
+@Global()
+@Module({
+    imports: [
+        TypeOrmModule.forFeature([ItemReadRepository, ItemWriteRepository]),
+    ]
+})
 export class InfrastructureModule {}
