+++ src/item/infrastructure/repositories/index.ts
@@ -1,0 +1,2 @@
+export * from './item-read.repository';
+export * from './item-write.repository';
