+++ src/item/infrastructure/repositories/item-read.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 
-import { ItemEntity } from '../entities';
+import { ItemEntity } from '../../entities';
 
 @EntityRepository(ItemEntity)
 export class ItemReadRepository extends Repository<ItemEntity> {}
