+++ src/item/repositories/index.ts
@@ -1,3 +1,1 @@
 export * from './item.repository';
-export * from './item-read.repository';
-export * from './item-write.repository';
